Move GeckoCodeDiag download logic to GeckoCodeConfig
