+++ Source/Core/Core/GeckoCodeConfig.h
@@ -4,6 +4,7 @@
 
 #pragma once
 
+#include <string>
 #include <vector>
 #include "Core/GeckoCode.h"
 
@@ -12,5 +13,6 @@ class IniFile;
 namespace Gecko
 {
 std::vector<GeckoCode> LoadCodes(const IniFile& globalIni, const IniFile& localIni);
+std::vector<GeckoCode> DownloadCodes(std::string gameid, bool* succeeded);
 void SaveCodes(IniFile& inifile, const std::vector<GeckoCode>& gcodes);
 }
